+++ CHANGELOG.md
@@ -19,7 +19,6 @@ All notable changes to this Android application project will be documented in th
 - [ ]  Update/Add javadocs comments.  Will be ongoing process to better document code.  
 - [ ]  Connectivity monitoring. 
 #### Changed
-- [ ]  Address TODO's relating to sendMessage().  
 #### Removed
 
 ## [Unreleased]
@@ -28,9 +27,13 @@ All notable changes to this Android application project will be documented in th
 - [x]  Update/Add javadocs comments.  Will be ongoing process to better document code.  
 - [x]  Create syslog and make it viewable.  The SSHFragment->txtReply is a poor man's version of syslog. 
 - [x]  Added GlobalVariables extends Application class.  This will contain shared preferences and other objects that are TBD.
-    - [ ]  Change code to utilize GlobalVariables.
+    - [x]  Change code to utilize GlobalVariables.
+- [ ]  **(STRETCH GOAL)**  System Status dialog.   Feature creep!  
+- [x]  System Logging.  Feature creep!
+   - [ ] Use across application.
 #### Changed
 - [ ]  SSH Fragment is a complete hack!  Replace or improve.
+- [x]  Address TODO's relating to EventBus->sendMessage().  
 #### Removed
 
 ## [Released]